+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- updating release action
+
 ## [13.0.9] - 2026-05-31
 
 - can now snooze a tab by right-clicking on the page