+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- added a release script to streamline the process
+
 ## [13.0.10] - 2026-06-01
 
 - updating release action